rename : 예약 관련 코드 reservation으로 이동
